+++ internal/domain/mocks/mock_notification_center_service.go
@@ -49,3 +49,17 @@ func (mr *MockNotificationCenterServiceMockRecorder) GetContactPreferences(arg0,
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetContactPreferences", reflect.TypeOf((*MockNotificationCenterService)(nil).GetContactPreferences), arg0, arg1, arg2, arg3)
 }
+
+// UpdateContactPreferences mocks base method.
+func (m *MockNotificationCenterService) UpdateContactPreferences(arg0 context.Context, arg1 *domain.UpdateContactPreferencesRequest) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "UpdateContactPreferences", arg0, arg1)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// UpdateContactPreferences indicates an expected call of UpdateContactPreferences.
+func (mr *MockNotificationCenterServiceMockRecorder) UpdateContactPreferences(arg0, arg1 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpdateContactPreferences", reflect.TypeOf((*MockNotificationCenterService)(nil).UpdateContactPreferences), arg0, arg1)
+}